+++ src/components/screens/vacation-requests-screen.tsx
@@ -249,7 +249,6 @@ const VacationRequestsScreen = () => {
 
     try {
       setLoading(true);
-      const days = calculateTotalVacationDays(vacationData.startDate, vacationData.endDate, getWorkingWeek(loggedInPerson));
       const createdRequest = await vacationRequestsApi.createVacationRequest({
         vacationRequest: {
           personId: userProfile.id,
@@ -260,7 +259,7 @@ const VacationRequestsScreen = () => {
           message: vacationData.message,
           createdAt: new Date(),
           updatedAt: new Date(),
-          days: days
+          days: vacationData.days
         }
       });
       setVacationRequests([createdRequest, ...vacationRequests]);
@@ -281,7 +280,6 @@ const VacationRequestsScreen = () => {
 
     try {
       setLoading(true);
-      const days = calculateTotalVacationDays(vacationData.startDate, vacationData.endDate, getWorkingWeek(loggedInPerson));
       const vacationRequest = vacationRequests.find(
         (vacationRequest) => vacationRequest.id === vacationRequestId
       );
@@ -295,7 +293,7 @@ const VacationRequestsScreen = () => {
             type: vacationData.type,
             message: vacationData.message,
             updatedAt: new Date(),
-            days: days
+            days: vacationData.days
           }
         });
         const updatedVacationRequests = vacationRequests.map((vacationRequest) =>
@@ -419,129 +417,6 @@ const VacationRequestsScreen = () => {
     );
     setLatestVacationRequestStatuses(createdAndFoundVacationRequestStatuses);
   };
-
-  /**
-   * Get a list of working days
-   *
-   * @param loggedInPerson person data
-   */
-  const getWorkingWeek = (loggedInPerson: any): boolean[] => {
-    const workingWeek = [false, false, false, false, false, false, false];
-    const daysOfWeek = ['monday', 'tuesday', 'wednesday', 'thursday', 'friday', 'saturday', 'sunday'];
-  
-    for (let i = 0; i < daysOfWeek.length; i++) {
-      if (loggedInPerson?.[daysOfWeek[i]] !== 0) {
-        workingWeek[i] = true;
-      }
-    }
-    return workingWeek;
-  }
-
-  /**
- * Calculates vacation days
- *
- * @param vacationDayStart vacation start date 
- * @param vacationDayEnd vacation end date
- * @param workingWeek list of working days
- */
-  const calculateTotalVacationDays = (vacationStartDate : DateTime, vacationEndDate : DateTime, workingWeek : boolean []) : number => {
-    let vacationDayStart = vacationStartDate.weekday;
-    let vacationDayEnd = vacationEndDate.weekday;
-    const diff = vacationEndDate.diff(vacationStartDate, ["days"]);
-    const days = Number(Math.round(diff.days));
-    let weeks = Math.floor(days/7);
-
-    const startWeek = getStartDayWorkingWeek(workingWeek);
-    const endWeek = getEndDayWorkingWeek(workingWeek);
-
-    if (startWeek === 0) return 0;
-
-    if (days === 0 && workingWeek[vacationDayStart-1]) {
-      if (endWeek - startWeek === 0) return 6;
-      return 1;
-    }
-    if (days === 0 && !workingWeek[vacationDayStart-1]) return 0;
-
-    if (vacationDayStart > endWeek && vacationDayEnd > endWeek || vacationDayEnd < startWeek && vacationDayStart < startWeek){
-      if (vacationDayEnd > vacationDayStart) return weeks * 6;
-      if (vacationDayEnd === vacationDayStart) return weeks * 6;
-      return weeks * 6 + 6;
-    }
-
-    if (vacationDayStart > endWeek && vacationDayEnd < startWeek) return weeks * 6;
-    
-    if (vacationDayEnd > endWeek) vacationDayEnd = endWeek;
-    if (vacationDayStart > endWeek) vacationDayStart = startWeek;
-    if (vacationDayEnd < startWeek) vacationDayEnd = endWeek;
-    if (vacationDayStart < startWeek) vacationDayStart = startWeek;
-      
-    if (vacationDayStart === vacationDayEnd){
-      if (endWeek - startWeek === 0) return weeks * 6 + 6;
-      if (vacationDayStart === startWeek) return weeks * 6 + 1;
-      if (vacationDayEnd === endWeek) return weeks * 6 + 1;
-      return weeks * 6;
-    }
-
-    if (vacationDayStart === startWeek){
-      if (vacationDayEnd === endWeek ) return weeks * 6 + 6;
-      let addDay = 0;
-      for (let i = vacationDayStart; i <= vacationDayEnd; i++){
-        if (workingWeek[i-1]) addDay+=1;
-      }
-      return weeks * 6 + addDay;
-    }
-    
-    if (vacationDayEnd > vacationDayStart){
-      let addDay = 0;
-      let startIndex = vacationDayStart;
-      if (weeks > 0) startIndex+=1;
-      if (vacationDayEnd === endWeek && weeks > 0 && workingWeek[vacationDayStart-1]) addDay += 1;
-      for (let i = startIndex; i <= vacationDayEnd; i++){
-        if (workingWeek[i-1]) addDay+=1;
-      }
-      return weeks * 6 + addDay;
-    }
-
-    let addDay = 0;
-    if (vacationDayEnd === vacationDayStart) weeks -= 1;
-    for (let i = vacationDayStart; i <= endWeek; i++){
-      if (workingWeek[i-1]) addDay+=1;
-    }
-    for (let i = startWeek; i <= vacationDayEnd; i++){
-      if (workingWeek[i-1]) addDay+=1;
-    }
-    return weeks * 6 + addDay;
-  }
-
-  /**
- * Get index - start day of working week
- *
- * @param workingWeek list of working days
- */
-  const getStartDayWorkingWeek = (workingWeek : boolean []) : number => {
-    let i = 0;
-    while (true) {
-      if (i>6) break;
-      if (workingWeek[i]) return i+1;
-      i++;
-    }
-    return 0;
-  }
-
-  /**
- * Get index - end day of working week
- *
- * @param workingWeek list of working days
- */
-  const getEndDayWorkingWeek = (workingWeek : boolean []) : number => {
-    let i = 6;
-    while (true) {
-      if (i<0) break;
-      if (workingWeek[i]) return i+1;
-      i--;
-    }
-    return 0;
-  }
 
   return (
     <>
